refactor(Parameters): move component to its own directory
