+++ src/components/menu/Parameters/Parameters.cy.tsx
