+++ src/components/sidebar/SideBar.tsx
@@ -3,7 +3,7 @@ import { Box, Drawer } from '@mui/material'
 
 import TransactionsMenu from '~components/menu/Transactions'
 import ReportsMenu from '~components/menu/Reports'
-import ParametersMenu from '~components/menu/Parameters'
+import ParametersMenu from '~/components/menu/Parameters/Parameters'
 import AnalysisMenu from '~/components/menu/Analysis/Analysis'
 
 export default function Sidebar() {
